Import/Export Writing Pipeline Utility
A new Pipeline option has been added to the menu. This is a space to develop features for internal usage in our production flow.

ConversationIO exports a ConversationManager node into a CSV file and imports a CSV file of the same structure to update dialogue and text of choices within the sequence. Game objects can be renamed for better clarity as well.